entity修正に伴う修正

+++ src/main/java/com/example/entity/Filmwork.java
@@ -44,7 +44,7 @@ public class Filmwork {
 	private String director;
 
 	@Column(name = "CATEGORY1_ID", nullable = true)
-	private Long category1Id;
+	private Integer category1Id;
 
     @ManyToOne
     @JoinColumn(name = "category1_id", insertable = false, updatable = false)
@@ -55,7 +55,7 @@ public class Filmwork {
     }
 
 	@Column(name = "CATEGORY2_ID", nullable = true)
-	private Long category2Id;
+	private Integer category2Id;
 
     @ManyToOne
     @JoinColumn(name = "category2_id", insertable = false, updatable = false)
@@ -66,7 +66,7 @@ public class Filmwork {
     }
 
 	@Column(name = "CATEGORY3_ID", nullable = true)
-	private Long category3Id;
+	private Integer category3Id;
 
     @ManyToOne
     @JoinColumn(name = "category3_id", insertable = false, updatable = false)
@@ -132,31 +132,31 @@ public class Filmwork {
 		this.director = director;
 	}
 
-	public Long getCategory1Id() {
+    public Integer getCategory1Id() {
 		return category1Id;
 	}
 
-	public void setCategory1Id(Long category1Id) {
+	public void setCategory1Id(Integer category1Id) {
 		this.category1Id = category1Id;
 	}
 
-	public Long getCategory2Id() {
+	public Integer getCategory2Id() {
 		return category2Id;
 	}
 
-	public void setCategory2Id(Long category2Id) {
+	public void setCategory2Id(Integer category2Id) {
 		this.category2Id = category2Id;
 	}
 
-	public Long getCategory3Id() {
+	public Integer getCategory3Id() {
 		return category3Id;
 	}
 
-	public void setCategory3Id(Long category3Id) {
+	public void setCategory3Id(Integer category3Id) {
 		this.category3Id = category3Id;
 	}
 
-    @OneToMany(mappedBy = "filmwork", cascade = CascadeType.ALL, fetch = FetchType.LAZY)
+	@OneToMany(mappedBy = "filmwork", cascade = CascadeType.ALL, fetch = FetchType.LAZY)
     private List<Review> reviews;
 
     public List<Review> getReviews() {
